Cleanup build script to current standards.

+++ scripts/build.sh
@@ -17,49 +17,55 @@
 #See the License for the specific language governing permissions and
 #limitations under the License.
 
+set -e
+set -u
+set -o pipefail
+
 GITVERSION=`git describe --always`
-SOURCEDIR=$1
+
+__dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+__proj_dir="$(dirname "$__dir")"
+
+SOURCEDIR="${__proj_dir}"
 BUILDDIR=$SOURCEDIR/build
 PLUGIN=`echo $SOURCEDIR | grep -oh "snap-plugin-.*"`
 ROOTFS=$BUILDDIR/rootfs
 BUILDCMD='go build -a -ldflags "-w"'
 
-echo
-echo "****  Prepare Heka  ****"
-echo
-OLDGOPATH=$GOPATH
-cd ../..
-GITHUBDIR=`pwd`
-mkdir -p mozilla-services
-cd mozilla-services
-if [ ! -d "heka" ] 
-then
-   git clone http://github.com/mozilla-services/heka
-fi
+# shellcheck source=scripts/common.sh
+. "${__dir}/common.sh"
 
-echo
-echo "****  Source Heka ENV  ****"
-echo
+mozilla_path=$(echo ${GOPATH//://src/github.com/mozilla-services:}/src/github.com/mozilla-services | cut -d':' -f1)
+heka_path="${mozilla_path}/heka"
 
-cd heka
-source ./build.sh
+_debug "heka path: ${heka_path}"
+_info "ensure latest mozilla heka repo available"
 
-# append old GOPATH to the current GOPATH
-export GOPATH=$GOPATH:$OLDGOPATH
+[[ -d ${heka_path} ]] || _fail "run 'make dep' and ensure mozilla/heka source exist in ${heka_path}"
 
-echo
-echo "****  snap Heka Plugin Build  ****"
-echo
-cd $SOURCEDIR
+_info "building heka"
+# NOTE: heka buid scripts does not honor set -e and set -u
+set +e
+set +u
+(cd "${heka_path}" && source ./build.sh)
+set -e
+set -u
+
+# append Heka build to GOPATH
+export GOPATH="${heka_path}/build/heka:${GOPATH}"
+_debug "heka custom GOPATH: ${GOPATH}"
+
+_info "building snap heka plugin"
+cd "${__proj_dir}"
 
 # Disable CGO for builds
 export CGO_ENABLED=0
 
 # Clean build bin dir
-rm -rf $ROOTFS/*
+rm -rf "${ROOTFS:?}"/*
 
 # Make dir
-mkdir -p $ROOTFS
+mkdir -p "${ROOTFS}"
 
 # Build plugin
 echo "Source Dir = $SOURCEDIR"
